Add command.
