+++ crates/cli/Cargo.toml
@@ -43,7 +43,7 @@ moon_vcs = { path = "../core/vcs" }
 moon_workspace = { path = "../core/workspace" }
 atty = "0.2.14"
 bytes = "1.4.0"
-clap = { workspace = true, features = ["derive", "env", "wrap_help"] }
+clap = { workspace = true }
 clap_complete = { workspace = true }
 clap_lex = "0.3.2"
 console = { workspace = true }
