+++ crates/cli/src/queries/mod.rs
@@ -1,3 +1,4 @@
+pub mod hash;
 pub mod hash_diff;
 pub mod projects;
 pub mod touched_files;
