+++ crates/core/action-context/Cargo.toml
@@ -5,6 +5,6 @@ edition = "2021"
 
 [dependencies]
 moon_target = { path = "../target" }
-clap = { workspace = true, features = ["derive"] }
+clap = { workspace = true }
 rustc-hash = { workspace = true }
 serde = { workspace = true }
